+++ test/libsolidity/smtCheckerTests/functions/abi_encode_functions.sol
@@ -6,5 +6,7 @@ contract C {
 }
 // ----
 // Warning: (31-64): Experimental features are turned on. Do not use experimental features on live deployments.
+// Warning: (162-165): Assertion checker does not yet implement type abi
 // Warning: (162-176): Assertion checker does not yet implement this type of function call.
+// Warning: (178-181): Assertion checker does not yet implement type abi
 // Warning: (178-203): Assertion checker does not yet implement this type of function call.