+++ test/libsolidity/smtCheckerTests/functions/functions_library_1_fail.sol
@@ -17,4 +17,5 @@ contract C
 	}
 }
 // ----
+// Warning: (228-229): Assertion checker does not yet implement type type(library L)
 // Warning: (245-261): Assertion violation happens here